WheatherApp/ViewModels/WheatherLatestViewModel.cs
using System;
using System.ComponentModel;
using WheatherApp.Domain.Repositories;
using WheatherApp.Infrastructure.SQLite;

namespace WheatherApp.ViewModels
{
    public class WheatherLatestViewModel : ViewModelBase
    {
        private IWheatherRepository _weather;

        public WheatherLatestViewModel() : this(new WheatherSQLite())
        {
        }

        public WheatherLatestViewModel(IWheatherRepository weather)
        {
            _weather = weather;
        }

        private string _areaIdText = string.Empty;
        public string AreaIdText
        {
            get { return _areaIdText; }
            set
            {
                SetProperty(ref _areaIdText, value);
            }
        }

        private string _dataDateText = string.Empty;
        public string DataDateText
        {
            get { return _dataDateText; }
            set
            {
                SetProperty(ref _dataDateText, value);
            }
        }

        private string _conditionText = string.Empty;
        public string ConditionText
        {
            get { return _conditionText; }
            set
            {
                SetProperty(ref _conditionText, value);
            }
        }

        private string _temperatureText = string.Empty;
        public string TemperatureText
        {
            get { return _temperatureText; }
            set
            {
                SetProperty(ref _temperatureText, value);
            }
        }

        public void Search()
        {
            var entity = _weather.GetLatest(Convert.ToInt32(AreaIdText));
            if (entity != null)
            {
                DataDateText = entity.DataDate.ToString();
                ConditionText = entity.Condition.DisplayValue;
                TemperatureText = entity.Temperature.DisplayValueWithUnitSpace;
            }
        }
    }
}
